backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pgvector==0.2.1
sentence-transformers==2.2.2
groq==0.4.0
openai==1.3.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0
redis==5.0.1
celery==5.3.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
